requirements.txt
streamlit==1.55.0
youtube-transcript-api==1.2.4
openai==2.28.0
python-dotenv==1.2.2
google-genai==1.67.0
yt-dlp==2026.3.13
pytest==8.4.2
